Move all server packages to server folder
